+++ lib/sqlite_httpvfs.dart
@@ -23,6 +23,7 @@ library;
 export 'src/constants.dart';
 export 'src/fetcher.dart';
 export 'src/fetcher_curl.dart';
+export 'src/fetcher_isolate.dart';
 export 'src/fetcher_socket.dart';
 export 'src/http_vfs.dart';
 export 'src/http_vfs_file.dart';